9043/refactor: extract processOrder methods
