+++ service-api/src/main/java/greencity/service/ubs/wayforpay/WayForPayService.java
@@ -4,6 +4,8 @@ import greencity.dto.order.PaymentSystemResponse;
 import greencity.dto.payment.PaymentWayForPayRequestDto;
 import greencity.entity.order.Order;
 
+//TODO add docs
+//TODO add test
 public interface WayForPayService {
     PaymentSystemResponse processWayForPay(Order order, long sumToPayInCoins);
