+++ service/src/main/java/greencity/service/ubs/UBSClientServiceImpl.java
@@ -11,35 +11,24 @@ import static greencity.constant.ErrorMessage.EMPLOYEE_DOESNT_EXIST;
 import static greencity.constant.ErrorMessage.EVENTS_NOT_FOUND_EXCEPTION;
 import static greencity.constant.ErrorMessage.LOCATION_DOESNT_FOUND_BY_ID;
 import static greencity.constant.ErrorMessage.LOCATION_IS_DEACTIVATED_FOR_TARIFF;
-import static greencity.constant.ErrorMessage.NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER;
-import static greencity.constant.ErrorMessage.ORDER_ALREADY_PAID;
-import static greencity.constant.ErrorMessage.ORDER_NOT_FOUND_BY_ID;
-import static greencity.constant.ErrorMessage.ORDER_STATUS_AND_PAYMENT_CONDITION_FAILED;
 import static greencity.constant.ErrorMessage.ORDER_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.RECIPIENT_WITH_CURRENT_ID_DOES_NOT_EXIST;
-import static greencity.constant.ErrorMessage.TARIFF_FOR_BAGS_AT_LOCATION_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_FOR_COURIER_AND_LOCATION_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_FOR_LOCATION_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_FOR_ORDER_NOT_EXIST;
 import static greencity.constant.ErrorMessage.TARIFF_NOT_FOUND;
 import static greencity.constant.ErrorMessage.TARIFF_NOT_FOUND_BY_LOCATION_ID;
 import static greencity.constant.ErrorMessage.TARIFF_OR_LOCATION_IS_DEACTIVATED;
-import static greencity.constant.ErrorMessage.TOO_MUCH_POINTS_FOR_ORDER;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_ID_DOES_NOT_EXIST;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_UUID_ALREADY_EXISTS_IN_UBS;
 import static greencity.constant.ErrorMessage.USER_WITH_CURRENT_UUID_DOES_NOT_EXIST;
-import static greencity.util.OrderUtils.getLastPayment;
 import static java.util.Objects.nonNull;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
-import com.google.maps.model.LatLng;
 import greencity.client.UserRemoteClient;
-import greencity.client.WayForPayClient;
 import greencity.constant.AppConstant;
 import greencity.constant.ErrorMessage;
-import greencity.constant.KyivTariffLocation;
 import greencity.constant.OrderHistory;
-import greencity.constant.TariffLocation;
 import greencity.dto.AllActiveLocationsDto;
 import greencity.dto.LocationWithTariffInfoDto;
 import greencity.dto.LocationsDto;
@@ -50,7 +39,6 @@ import greencity.dto.TariffInfoDto;
 import greencity.dto.TariffsForLocationDto;
 import greencity.dto.address.AddressDto;
 import greencity.dto.address.AddressInfoDto;
-import greencity.dto.bag.BagDto;
 import greencity.dto.bag.BagForUserDto;
 import greencity.dto.bag.BagTranslationDto;
 import greencity.dto.certificate.CertificateDto;
@@ -63,12 +51,8 @@ import greencity.dto.order.EventDto;
 import greencity.dto.order.OrderAddressDtoRequest;
 import greencity.dto.order.OrderCancellationReasonDto;
 import greencity.dto.order.OrderPaymentDetailDto;
-import greencity.dto.order.OrderResponseDto;
-import greencity.dto.order.OrderWayForPayClientDto;
 import greencity.dto.order.OrdersDataForUserDto;
-import greencity.dto.order.PaymentSystemResponse;
 import greencity.dto.pageble.PageableDto;
-import greencity.dto.payment.PaymentWayForPayRequestDto;
 import greencity.dto.position.PositionAuthoritiesDto;
 import greencity.dto.user.AllPointsUserDto;
 import greencity.dto.user.DeactivateUserRequestDto;
@@ -80,13 +64,11 @@ import greencity.dto.user.UserPointsAndAllBagsDto;
 import greencity.dto.user.UserProfileCreateDto;
 import greencity.dto.user.UserProfileDto;
 import greencity.dto.user.UserProfileUpdateDto;
-import greencity.entity.coords.Coordinates;
 import greencity.entity.order.Bag;
 import greencity.entity.order.Certificate;
 import greencity.entity.order.ChangeOfPoints;
 import greencity.entity.order.Event;
 import greencity.entity.order.Order;
-import greencity.entity.order.OrderBag;
 import greencity.entity.order.OrderPaymentStatusTranslation;
 import greencity.entity.order.OrderStatusTranslation;
 import greencity.entity.order.Payment;
@@ -98,18 +80,13 @@ import greencity.entity.user.employee.Employee;
 import greencity.entity.user.ubs.Address;
 import greencity.entity.user.ubs.OrderAddress;
 import greencity.entity.user.ubs.UBSuser;
-import greencity.enums.AddressStatus;
-import greencity.enums.BonusReason;
 import greencity.enums.BotType;
 import greencity.enums.CertificateStatus;
 import greencity.enums.LocationStatus;
-import greencity.enums.OrderPaymentStatus;
 import greencity.enums.OrderStatus;
-import greencity.enums.PaymentStatus;
 import greencity.enums.TariffStatus;
 import greencity.exceptions.BadRequestException;
 import greencity.exceptions.NotFoundException;
-import greencity.exceptions.address.AddressNotWithinLocationAreaException;
 import greencity.exceptions.http.AccessDeniedException;
 import greencity.exceptions.user.UBSuserNotFoundException;
 import greencity.exceptions.user.UserNotFoundException;
@@ -121,7 +98,6 @@ import greencity.repository.CourierRepository;
 import greencity.repository.EmployeeRepository;
 import greencity.repository.EventRepository;
 import greencity.repository.LocationRepository;
-import greencity.repository.OrderAddressRepository;
 import greencity.repository.OrderBagRepository;
 import greencity.repository.OrderPaymentStatusTranslationRepository;
 import greencity.repository.OrderRepository;
@@ -132,33 +108,21 @@ import greencity.repository.TariffsInfoRepository;
 import greencity.repository.TelegramChatRepository;
 import greencity.repository.UBSUserRepository;
 import greencity.repository.UserRepository;
-import greencity.service.DistanceCalculationUtils;
-import greencity.service.google.GoogleApiService;
 import greencity.service.phone.UAPhoneNumberUtil;
 import greencity.service.ubs.calculator.BagCalculatorService;
 import greencity.service.ubs.calculator.CertificateCalculatorService;
 import greencity.service.ubs.calculator.PaymentCalculatorService;
-import greencity.service.ubs.calculator.PointCalculatorService;
-import greencity.service.ubs.payment.PaymentStrategyFactory;
-import greencity.service.ubs.wayforpay.WayForPayService;
 import greencity.util.Bot;
 import greencity.util.MoneyConverterUtil;
-import greencity.util.OrderUtils;
-import greencity.util.PointsUtils;
 import jakarta.transaction.Transactional;
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.EnumSet;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -168,7 +132,6 @@ import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections4.CollectionUtils;
 import org.json.JSONObject;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Value;
@@ -195,31 +158,23 @@ public class UBSClientServiceImpl implements UBSClientService {
     private final CourierRepository courierRepository;
     private final EmployeeRepository employeeRepository;
     private final AddressRepository addressRepo;
-    private final OrderAddressRepository orderAddressRepository;
     private final UserRemoteClient userRemoteClient;
     private final EventRepository eventRepository;
     private final OrdersForUserRepository ordersForUserRepository;
     private final OrderStatusTranslationRepository orderStatusTranslationRepository;
     private final OrderPaymentStatusTranslationRepository orderPaymentStatusTranslationRepository;
-    private final GoogleApiService googleApiService;
     private final EventService eventService;
     private final TariffLocationRepository tariffLocationRepository;
     private final LocationRepository locationRepository;
     private final TariffsInfoRepository tariffsInfoRepository;
     private final TelegramChatRepository telegramBotRepository;
     private final OrderBagRepository orderBagRepository;
-    private final NotificationService notificationService;
-    private final WayForPayClient wayForPayClient;
     private final LocationToLocationsDtoMapper locationToLocationsDtoMapper;
     private final AddressService addressService;
-    private final PointsUtils pointsUtils;
     private final PaymentCalculatorService paymentCalculatorService;
     private final BagCalculatorService bagCalculatorService;
-    private final PointCalculatorService pointCalculatorService;
     private final CertificateCalculatorService certificateCalculatorService;
     private final MoneyConverterUtil moneyConverterUtil;
-    private final PaymentStrategyFactory paymentStrategyFactory;
-    private final WayForPayService wayForPayService;
 
     @Value("${greencity.bots.ubs-bot-name}")
     private String telegramBotName;
@@ -293,7 +248,6 @@ public class UBSClientServiceImpl implements UBSClientService {
             orderId);
     }
 
-    //TODO make duplicate or move to conroler like in project learning with pre authorised
     private void checkIsOrderOfCurrentUser(User user, Order order) {
         if (!order.getUser().getId().equals(user.getId())) {
             throw new AccessDeniedException(ErrorMessage.ORDER_DOES_NOT_BELONG_TO_USER);
@@ -402,250 +356,6 @@ public class UBSClientServiceImpl implements UBSClientService {
                 .build();
         }
         return modelMapper.map(certificate, CertificateDto.class);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @Transactional
-    public PaymentSystemResponse processNewOrder(OrderResponseDto dto, String uuid) {
-        validateOrderRequestAddress(dto);//todo move
-
-        adjustPaymentDetails(dto);//todo move
-
-        Order order = modelMapper.map(dto, Order.class);
-        order.setOrderDate(LocalDateTime.now());
-        order.setOrderStatus(OrderStatus.FORMED);
-        order.setCounterOrderPaymentId(0L);
-
-        User currentUser = userRepository.findByUuid(uuid);
-
-        //todo move
-        OrderAddress orderAddress = formAndSaveOrderAddress(
-            dto.getAddressId(), dto.getLocationId(), currentUser);
-
-        //todo move
-        UBSuser userData = formAndSaveUbsUser(
-            dto.getPersonalData(), null, orderAddress, currentUser);
-        //TODO another changes think about OrderService
-        order = formAndSaveOrderRequest(dto, order, currentUser, userData);
-        long sumToPayInCoins = getLastPayment(order).getAmount();
-
-        //todo move
-        formAndSaveUser(currentUser, dto.getPointsToUse(), order);
-
-        //todo move
-        saveOrderEvent(OrderHistory.ORDER_FORMED_UK, OrderHistory.CLIENT_UK, order);
-
-        //todo move
-        PaymentSystemResponse paymentSystemResponse =
-            processPaymentResponse(dto, order, sumToPayInCoins);
-
-        notificationService.notifyCreatedOrder(order);
-
-        return paymentSystemResponse;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @Transactional
-    public PaymentSystemResponse processExistingOrder(OrderResponseDto dto, String uuid, Long orderId) {
-        validateOrderRequestAddress(dto);//todo move
-
-        Order order = orderRepository.findById(orderId)
-            .orElseThrow(() -> new NotFoundException(ORDER_NOT_FOUND_BY_ID + orderId));
-
-        User currentUser = userRepository.findByUuid(uuid);
-        checkIsOrderOfCurrentUser(currentUser, order);// todo move
-
-        if (order.getOrderStatus() != OrderStatus.FORMED
-            || order.getOrderPaymentStatus() != OrderPaymentStatus.UNPAID) {
-            throw new BadRequestException(ORDER_STATUS_AND_PAYMENT_CONDITION_FAILED);
-        }
-
-        adjustPaymentDetails(dto);//todo move
-
-        order.setPointsToUse(dto.getPointsToUse());
-        order.setAdditionalOrders(dto.getAdditionalOrders());
-        order.setComment(dto.getOrderComment());
-
-        //todo move
-        OrderAddress orderAddress = getOrUpdateOrderAddress(
-            order.getUbsUser().getOrderAddress(), dto.getAddressId(), dto.getLocationId(), currentUser);
-
-        //todo move
-        UBSuser userData = formAndSaveUbsUser(
-            dto.getPersonalData(), order.getUbsUser().getId(), orderAddress, currentUser);
-        //TODO another changes think about OrderService
-        order = formAndSaveOrderRequest(dto, order, currentUser, userData);
-        long sumToPayInCoins = getLastPayment(order).getAmount();
-
-        //todo move
-        formAndSaveUser(currentUser, dto.getPointsToUse(), order);
-
-        //todo move
-        saveOrderEvent(OrderHistory.ORDER_STATUS_UPDATED_UK, OrderHistory.CLIENT_UK, order);
-
-        //todo move
-        PaymentSystemResponse paymentSystemResponse =
-            processPaymentResponse(dto, order, sumToPayInCoins);
-
-        if (order.getOrderPaymentStatus() == OrderPaymentStatus.UNPAID) {
-            notificationService.notifyUnpaidOrderPermanently(order, sumToPayInCoins, paymentSystemResponse);
-        }
-
-        return paymentSystemResponse;
-    }
-
-    //TODO move to ProcessOrder (processNewOrder, processExistingOrder)
-    private void validateOrderRequestAddress(OrderResponseDto dto) {
-        if (!checkIfAddressMatchLocationArea(dto.getLocationId(), dto.getAddressId())) {
-            throw new AddressNotWithinLocationAreaException(AppConstant.ADDRESS_NOT_WITHIN_LOCATION_AREA_MESSAGE);
-        }
-    }
-
-    //TODO move to ProcessOrder (processNewOrder, processExistingOrder)
-    private void adjustPaymentDetails(OrderResponseDto dto) {
-        if (!dto.isShouldBePaid()) {
-            dto.setCertificates(Collections.emptySet());
-            dto.setPointsToUse(0);
-        }
-    }
-
-    //TODO move first part to calculator somewhere and move all method to OrderService
-    //TODO think to use some inner class or dto
-    //TODO or just move all method to OrderService
-    private Order formAndSaveOrderRequest(OrderResponseDto dto, Order order, User currentUser, UBSuser userData) {
-        TariffsInfo tariffsInfo = findTariffsInfoByBagIdsWithinLocation(
-            getBagIds(dto.getBags()), dto.getLocationId());//todo move
-        List<OrderBag> bagsOrdered = new ArrayList<>();
-        long sumToPayInCoinsWithoutDiscount =
-            bagCalculatorService.prepareBagsAndCalculateTotal(
-                bagsOrdered, dto.getBags(), tariffsInfo);
-
-        pointsUtils.checkIfUserHaveEnoughPoints(currentUser.getCurrentPoints(), dto.getPointsToUse());
-        long sumToPayInCoins = pointCalculatorService
-            .reduceOrderSumDueToUsedPoints(sumToPayInCoinsWithoutDiscount, dto.getPointsToUse());
-        if (sumToPayInCoinsWithoutDiscount == sumToPayInCoins) {
-            order.setPointsToUse(0);
-            dto.setPointsToUse(0);
-        }
-
-        Set<Certificate> orderCertificates = new HashSet<>();
-        sumToPayInCoins =
-            certificateCalculatorService.applyCertificatesToOrder(
-                dto, orderCertificates, order, sumToPayInCoins);
-        if (sumToPayInCoins <= 0) {
-            dto.setShouldBePaid(false);
-        }
-        //todo move
-        return formAndSaveOrder(order, orderCertificates, bagsOrdered, userData, currentUser, sumToPayInCoins,
-            tariffsInfo);
-    }
-
-    //TODO move to EventService (processNewOrder, processExistingOrder)
-    private void saveOrderEvent(String eventName, String author, Order order) {
-        eventService.save(eventName, author, order);
-        log.info("Saved event: eventName={}, author={}, orderId={}", eventName, author, order.getId());
-    }
-
-    //TODO move to ProcessOrder (processNewOrder, processExistingOrder)
-    private PaymentSystemResponse processPaymentResponse(OrderResponseDto dto, Order order, long sumToPayInCoins) {
-        if (dto.isShouldBePaid()) {
-            return paymentStrategyFactory.getPaymentStrategy(dto.getPaymentSystem())
-                .processPayment(order, sumToPayInCoins);
-        } else {
-            return wayForPayService.getPaymentRequestDto(order, "");
-        }
-    }
-
-    private boolean checkIfAddressMatchLocationArea(long locationId, long addressId) {
-        Address address = addressRepo.findById(addressId)
-            .orElseThrow(() -> new NotFoundException(AppConstant.ADDRESS_NOT_FOUND_BY_ID_MESSAGE + addressId));
-
-        boolean isKyivTariff = checkIfCityBelongsToKyivTariff(address.getBaseAddress().getCityEn());
-
-        if (locationId == TariffLocation.KYIV_TARIFF.getLocationId()) {
-            return isKyivTariff;
-        } else if (locationId == TariffLocation.KYIV_REGION_20_KM_TARIFF.getLocationId()) {
-            checkAndCalculateAddressCoordinatesIfEmpty(address);
-
-            double addressLatitude = address.getCoordinates().getLatitude();
-            double addressLongitude = address.getCoordinates().getLongitude();
-
-            double distanceInKm =
-                DistanceCalculationUtils.calculateDistanceInKmByHaversineFormula(AppConstant.KYIV_LATITUDE,
-                    AppConstant.KYIV_LONGITUDE,
-                    addressLatitude, addressLongitude);
-
-            return distanceInKm <= AppConstant.LOCATION_40_KM_ZONE_VALUE && !isKyivTariff;
-        } else {
-            return locationRepository.findAddressAndLocationNamesMatch(locationId, addressId).isPresent();
-        }
-    }
-
-    private boolean checkIfCityBelongsToKyivTariff(String cityName) {
-        return Arrays.stream(KyivTariffLocation.values())
-            .anyMatch(kyivTariffLocation -> kyivTariffLocation.getLocationName().equalsIgnoreCase(cityName));
-    }
-
-    //TODO think where to move
-    private void checkAndCalculateAddressCoordinatesIfEmpty(Address address) {
-        if (address.getCoordinates().getLatitude() == 0.0 && address.getCoordinates().getLongitude() == 0.0) {
-            LatLng latLng = googleApiService
-                .getGeocodingResultByCityAndCountryAndLocale(AppConstant.UKRAINE_EN,
-                    address.getBaseAddress().getCityEn(),
-                    AppConstant.LANG_EN).geometry.location;
-            Coordinates addressCoordinates = Coordinates.builder().latitude(latLng.lat).longitude(latLng.lng).build();
-            address.setCoordinates(addressCoordinates);
-            addressRepo.save(address);
-        }
-    }
-
-    private List<Integer> getBagIds(List<BagDto> dto) {
-        return dto.stream()
-            .map(BagDto::getId)
-            .toList();
-    }
-
-    //TODO move to tarifs
-    private TariffsInfo findTariffsInfoByBagIdsWithinLocation(List<Integer> bagIds, Long locationId) {
-        return tariffsInfoRepository.findTariffsInfoByBagIdAndLocationId(bagIds, locationId)
-            .orElseThrow(
-                () -> new NotFoundException(String.format(TARIFF_FOR_BAGS_AT_LOCATION_NOT_EXIST, bagIds, locationId)));
-    }
-
-    private void checkIfAddressHasBeenDeleted(Address address) {
-        if (address.getBaseAddress().getAddressStatus().equals(AddressStatus.DELETED)) {
-            throw new NotFoundException(
-                NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + address.getId());
-        }
-    }
-
-    private void checkAddressUser(Address address, User user) {
-        if (!address.getUser().equals(user)) {
-            throw new NotFoundException(
-                NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + address.getId());
-        }
-    }
-
-    //TODO move to ProcessOrder (processNewOrder, processExistingOrder)
-    private void formAndSaveUser(User currentUser, int pointsToUse, Order order) {
-        currentUser.getOrders().add(order);
-        if (pointsToUse != 0) {
-            currentUser.setCurrentPoints(currentUser.getCurrentPoints() - pointsToUse);
-            currentUser.getChangeOfPointsList().add(ChangeOfPoints.builder()
-                .amount(-pointsToUse)
-                .date(order.getOrderDate())
-                .user(currentUser)
-                .order(order)
-                .reason(BonusReason.DEBIT_PAYMENT)
-                .build());
-        }
-        userRepository.save(currentUser);
     }
 
     /**
@@ -899,78 +609,6 @@ public class UBSClientServiceImpl implements UBSClientService {
         return ubsUser;
     }
 
-    //TODO move to OrderService
-    private Order formAndSaveOrder(
-        Order order, Set<Certificate> orderCertificates, List<OrderBag> bagsOrdered,
-        UBSuser userData, User currentUser, long sumToPayInCoins, TariffsInfo tariffsInfo
-    ) {
-        order.setTariffsInfo(tariffsInfo);
-        order.setCertificates(orderCertificates);
-        order.setOrderBags(bagsOrdered);
-        order.setUbsUser(userData);
-        order.setUser(currentUser);
-        order.setSumTotalAmountWithoutDiscounts(
-            paymentCalculatorService.calculateOrderSumWithoutDiscounts(bagsOrdered));
-        order.setCounterOrderPaymentId(order.getCounterOrderPaymentId() + 1);
-        setOrderPaymentStatus(order, sumToPayInCoins);
-
-        Payment payment = Payment.builder()
-            .amount(sumToPayInCoins)
-            .orderStatus(OrderStatus.FORMED)
-            .currency("UAH")
-            .paymentStatus(PaymentStatus.UNPAID)
-            .settlementDate(LocalDate.now().format(DateTimeFormatter.ISO_LOCAL_DATE))
-            .order(order).build();
-
-        if (order.getPayment() == null) {
-            order.setPayment(new ArrayList<>());
-        }
-        order.getPayment().add(payment);
-        return orderRepository.save(order);
-    }
-
-    private void setOrderPaymentStatus(Order order, long sumToPay) {
-        if (sumToPay <= 0) {
-            order.setOrderPaymentStatus(OrderPaymentStatus.PAID);
-        } else {
-            order.setOrderPaymentStatus(
-                order.getPointsToUse() > 0 || CollectionUtils.isNotEmpty(order.getCertificates())
-                    ? OrderPaymentStatus.HALF_PAID
-                    : OrderPaymentStatus.UNPAID);
-        }
-    }
-
-    //TODO move to UserService (processNewOrder, processExistingOrder)
-    private UBSuser formAndSaveUbsUser(
-        PersonalDataDto dto, Long id, OrderAddress orderAddress, User currentUser) {
-        UBSuser userData = modelMapper.map(dto, UBSuser.class);
-        userData.setId(id);
-        userData.setUser(currentUser);
-        userData.setPhoneNumber(
-            UAPhoneNumberUtil.getE164PhoneNumberFormat(userData.getPhoneNumber()));
-        userData.setOrderAddress(orderAddress);
-        userData = ubsUserRepository.save(userData);
-
-        currentUser.getUbsUsers().add(userData);
-        currentUser.setRecipientSurname(dto.getLastName());
-        currentUser.setRecipientName(dto.getFirstName());
-        currentUser.setRecipientPhone(dto.getPhoneNumber());
-        userRepository.save(currentUser);
-
-        return userData;
-    }
-
-    //TODO move to AdressService (processExistingOrder)
-    private OrderAddress getOrUpdateOrderAddress(
-        OrderAddress currentOrderAddress, Long newAddressId, Long newLocationId, User currentUser) {
-        OrderAddress newOrderAddress = formOrderAddress(newAddressId, newLocationId, currentUser);
-        newOrderAddress.setId(currentOrderAddress.getId());
-        if (currentOrderAddress.equals(newOrderAddress)) {
-            return currentOrderAddress;
-        }
-        return orderAddressRepository.save(newOrderAddress);
-    }
-
     /**
      * {@inheritDoc}
      */
@@ -1135,26 +773,6 @@ public class UBSClientServiceImpl implements UBSClientService {
             .build();
     }
 
-    //TODO move to OrderAdress (processNewOrder)
-    private OrderAddress formAndSaveOrderAddress(Long addressId, Long locationId, User currentUser) {
-        return orderAddressRepository.save(formOrderAddress(addressId, locationId, currentUser));
-    }
-
-    private OrderAddress formOrderAddress(Long addressId, Long locationId, User currentUser) {
-        Address address = addressRepo.findById(addressId)
-            .orElseThrow(() -> new NotFoundException(NOT_FOUND_ADDRESS_ID_FOR_CURRENT_USER + addressId));
-        Location location = locationRepository.findById(locationId)
-            .orElseThrow(() -> new NotFoundException(LOCATION_DOESNT_FOUND_BY_ID + locationId));
-
-        checkIfAddressHasBeenDeleted(address);
-        checkAddressUser(address, currentUser);
-
-        OrderAddress orderAddress = modelMapper.map(address, OrderAddress.class);
-        orderAddress.setLocation(location);
-
-        return orderAddress;
-    }
-
     @Override
     @Transactional
     public void deleteOrder(String uuid, Long id) {
@@ -1300,113 +918,6 @@ public class UBSClientServiceImpl implements UBSClientService {
     @Override
     public void updateEmployeesAuthorities(UserEmployeeAuthorityDto dto) {
         userRemoteClient.updateEmployeesAuthorities(dto);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public PaymentSystemResponse processOrder(String userUuid, OrderWayForPayClientDto dto) {
-        Order order = orderRepository.findById(dto.getOrderId())
-            .orElseThrow(() -> new NotFoundException(ORDER_WITH_CURRENT_ID_DOES_NOT_EXIST + dto.getOrderId()));
-        checkOrderIsPaid(order.getOrderPaymentStatus());
-        User currentUser = userRepository.findUserByUuid(userUuid)
-            .orElseThrow(() -> new NotFoundException(USER_WITH_CURRENT_UUID_DOES_NOT_EXIST + userUuid));
-        checkForNullCounter(order);
-        long sumToPayInCoins = paymentCalculatorService.calculateSumToPay(dto, order, currentUser);
-
-        transferUserPointsToOrder(order, dto.getPointsToUse());//todo move
-        paymentVerification(sumToPayInCoins, order);//todo move
-
-        if (sumToPayInCoins <= 0) {
-            return wayForPayService.getPaymentRequestDto(order, null);
-        } else {
-            String link = formedLink(order, sumToPayInCoins);//todo move
-            return wayForPayService.getPaymentRequestDto(order, link);
-        }
-    }
-
-    //TODO move (processOrder)
-    @Override
-    public String formedLink(Order order, long sumToPayInCoins) {
-        Order increment = incrementCounter(order);//todo move
-        PaymentWayForPayRequestDto paymentWayForPayRequestDto =
-            wayForPayService.formPaymentRequestForWayForPay(increment.getId(), sumToPayInCoins);
-        paymentWayForPayRequestDto
-            .setOrderReference(OrderUtils.generateEncodedOrderReference(increment.getId(), order));
-        return wayForPayService.getLinkFromWayForPayCheckoutResponse(
-            wayForPayClient.getCheckOutResponse(paymentWayForPayRequestDto));
-    }
-
-    //TODO move (processOrder)
-    private Order incrementCounter(Order order) {
-        order.setCounterOrderPaymentId(order.getCounterOrderPaymentId() + 1);
-        orderRepository.save(order);
-        return order;
-    }
-
-    //TODO move (processOrder)
-    private void paymentVerification(long sumToPayInCoins, Order order) {
-        if (sumToPayInCoins <= 0) {
-            order.setOrderPaymentStatus(OrderPaymentStatus.PAID);
-            order.setOrderStatus(OrderStatus.CONFIRMED);
-            orderRepository.save(order);
-            eventService.save(OrderHistory.ORDER_CONFIRMED_UK, OrderHistory.SYSTEM_UK, order);
-        }
-    }
-
-    //TODO move (processOrder)
-    private void transferUserPointsToOrder(Order order, Integer pointsToUse) {
-        if (pointsToUse <= 0) {
-            return;
-        }
-
-        User user = order.getUser();
-        pointsUtils.checkIfUserHaveEnoughPoints(user.getCurrentPoints(), pointsToUse);
-
-        int maxPointsToTransfer = countAmountToPayForOrder(order); //todo move
-        if (pointsToUse > maxPointsToTransfer) {
-            throw new BadRequestException(TOO_MUCH_POINTS_FOR_ORDER + maxPointsToTransfer);
-        }
-
-        order.setPointsToUse(order.getPointsToUse() + pointsToUse);
-        user.setCurrentPoints(user.getCurrentPoints() - pointsToUse);
-        user.getChangeOfPointsList()
-            .add(ChangeOfPoints.builder()
-                .user(user)
-                .amount(-pointsToUse)
-                .date(LocalDateTime.now())
-                .order(order)
-                .reason(BonusReason.DEBIT_PAYMENT)
-                .build());
-
-        orderRepository.save(order);
-    }
-
-    //TODO move (processOrder) (move somewhere to calculating)
-    private int countAmountToPayForOrder(Order order) {
-        int certificatesAmount = nonNull(order.getCertificates())
-            ? order.getCertificates().stream()
-                .map(Certificate::getPoints)
-                .reduce(0, Integer::sum)
-            : 0;
-        return -order.getPointsToUse() - certificatesAmount
-            + BigDecimal.valueOf(order.getSumTotalAmountWithoutDiscounts())
-                .movePointLeft(AppConstant.TWO_DECIMALS_AFTER_POINT_IN_CURRENCY)
-                .setScale(0, RoundingMode.UP).intValue();
-    }
-
-    //TODO move for (processOrder)
-    private void checkOrderIsPaid(OrderPaymentStatus orderPaymentStatus) {
-        if (OrderPaymentStatus.PAID.equals(orderPaymentStatus)) {
-            throw new BadRequestException(ORDER_ALREADY_PAID);
-        }
-    }
-
-    private void checkForNullCounter(Order order) {
-        if (order.getCounterOrderPaymentId() == null) {
-            order.setCounterOrderPaymentId(0L);
-        }
     }
 
     /**
